Fix 1.1 support, move more events and methods
