+++ src/main/java/net/serble/estools/ServerApi/Implementations/Bukkit/EventHandlers/BukkitEventsListenerPost1_4.java
@@ -14,7 +14,7 @@ import org.bukkit.event.inventory.InventoryDragEvent;
 
 import java.util.Set;
 
-public class BukkitEventsListenerPost1_4 extends BukkitEventsListener {
+public class BukkitEventsListenerPost1_4 extends BukkitEventsListenerPost1_1 {
 
     @EventHandler
     public void onDrag(InventoryDragEvent e) {
